requirements.txt
pdfplumber>=0.10.0
PyPDF2>=3.0.0
google-genai>=1.0.0
streamlit>=1.30.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
